Changing `eventgrid.App` to `buffalo.App`
Previously, `eventgrid.App` was a struct composed of a `buffalo.App`.
The idea being that one could have an eventgrid.App that exposed all of
the functionality seamlessly of a `buffalo.App`. However, in practice,
everybody will construct a `buffalo.App` first, because that's the
primary piece of functionality they're looking for. The process to build
an EventGrid app from a buffalo app was a huge pain. For that reason,
changing it so that one simply casts is a much easier way to consume it.