Removing "flavor" flag.
Before this project was going to be a buffalo plugin, it was going to be
our own rapid-devlopment code generator. At that point, it made sense to
generate code in support of many different HTTP frameworks. That
doesn't make any sense anymore, so I'm removing the code artifacts that
suggest that strategy.