Removing SubscriptionValidation default binding.
Before this change, `TypeDispatchSubscriber.Receive` would forward the `buffalo.Context`
directly. This gave the `ReceiveSubscriptionValidationRequest` function the ability to
respond to the request with some special sauce. Now that we discard anything written to
the context by wrapping it with our own type, that response will no longer get logged.

Inorder to fix this, I've removed the automatic binding of the
SubscriptionValidationResponse, and instead expose it as Middleware.