Parallelize Event Batch Processing (#12)
* Adding more Mock HTTP implementation.

* Formatting

* Fixing broken closure, as caught by `go vet`.

* Adding "-race" flag to CI tests.

Now that eventgrid processes batches in parallel, it makes sense to
sanity check our implementation for race conditions.

* Fixing `golint` violations.

* Removing SubscriptionValidation default binding.

Before this change, `TypeDispatchSubscriber.Receive` would forward the `buffalo.Context`
directly. This gave the `ReceiveSubscriptionValidationRequest` function the ability to
respond to the request with some special sauce. Now that we discard anything written to
the context by wrapping it with our own type, that response will no longer get logged.

Inorder to fix this, I've removed the automatic binding of the
SubscriptionValidationResponse, and instead expose it as Middleware.

* Refactoring SubscriptionValidation into Middleware

* fixing up some comments